Enhanced documentation
